Add flags for publish (keep draft/timestamp)
This adds command line flags to keep the draft file or to not update the
timestamp should they be provided

Conflicts:
	lib/mr_poole/version.rb

+++ lib/mr_poole/commands.rb
@@ -66,7 +66,12 @@ module MrPoole
     end
 
     # Todo make this take a path instead?
-    def publish(draftpath)
+    #
+    # @param draftpath [String] path to the draft
+    # @option options :keep_draft [Boolean] if true, keep the draft file
+    # @option options :keep_timestamp [Boolean] if true, don't change the timestamp
+    def publish(draftpath, opts = {})
+      opts = @helper.ensure_open_struct(opts)
       tail = File.basename(draftpath)
 
       begin
@@ -82,13 +87,13 @@ module MrPoole
       outfile = File.open(outpath, "w")
 
       infile.each_line do |line|
-        l = line.sub(/^date:\s*$/, "date: #{date} #{time}\n")
-        outfile.write(l)
+        line.sub!(/^date:\s*$/, "date: #{date} #{time}\n") unless opts.keep_timestamp
+        outfile.write(line)
       end
 
       infile.close
       outfile.close
-      FileUtils.rm(draftpath)
+      FileUtils.rm(draftpath) unless opts.keep_draft
 
       outpath
     end
